refactor: utils로 함수 책임 분리
VC 생성, 타임스탬프 코드를 Service 로직과 분리
